requirements.txt
unstructured
tiktoken
pinecone
langchain
pandas 
numpy
python-dotenv
langchain-community
flask
langgraph
pymongo
psycopg2
langchain-groq